test(atb1): pin the bool() wrapper with a numpy input; honest test-evidence wording
Both core reviewers passed (z-code-reviewer APPROVED, z-spec-compliance-review
PASS). Two non-blocking nits, both addressed here; no production logic changed.

- LOW (code-reviewer): test_returns_plain_bool_not_numpy used float literals, so
  `type(out) is bool` passed even without the bool() wrapper — it did not
  reproduce the numpy.float64 production input the wrapper defends against. Now
  feeds np.float64(0.80); verified it goes red if the wrapper is removed.

- WARNING (spec): the "all 5 fail pre-fix" claim was literally true but the red
  is an AttributeError (helper absent pre-fix), not a behavioral measurement of
  the insert bug. CHANGELOG and ledger now say so plainly, matching the honesty
  bar the P1.44 entry set, and note the insert call-site is covered only
  indirectly (run_trendline_detector is not drivable).

Refs: T-2026-CU-9050-062, AUDIT_TODO P1.47

+++ backtest/test_atb1_posted_flag.py
@@ -22,6 +22,8 @@ import os
 import sys
 import unittest.mock as mock
 
+import numpy as np
+
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 # core.config fails hard on missing secrets; the build machine ships an empty .env.
@@ -88,6 +90,11 @@ def test_boundary_is_not_the_025_shadow_gate():
 
 
 def test_returns_plain_bool_not_numpy():
-    # psycopg2 wants a real bool for the posted column, not numpy.bool_
-    out = atb1._atb1_posted_flag(0.80, 0.70)
+    # psycopg2 wants a real bool for the posted column, not numpy.bool_.
+    # ml_prob comes from model.predict_proba, i.e. numpy.float64 in production —
+    # numpy.float64 >= float yields numpy.bool_, so the helper's bool() wrapper
+    # is load-bearing. Feed a numpy input so this test actually pins that: with a
+    # plain float literal, `>=` already returns a Python bool and the assertion
+    # would pass even if the wrapper were deleted.
+    out = atb1._atb1_posted_flag(np.float64(0.80), 0.70)
     assert type(out) is bool
